src/requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
flask_pymongo==2.3.0
flask_wtf==0.14.3
python_bcrypt==0.3.2
WTForms==2.3.1
